+++ frontend/src/pages/Support.js
@@ -3,7 +3,7 @@ import React from "react";
 //******* Import component(s) *******/
 import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
-
+import "../styles/CSS/support.css";
 //******* Import style logic *******/
 //Here
 
@@ -12,9 +12,155 @@ export default function Support() {
     <div>
       <Navbar />
       <div className="container">
-        <h1>Support page</h1>
-        <h2>Here is your support plateform, we are here to help you!</h2>
+        <h1>Support page</h1>    <br/>
+        <h2>Welcome to the StoryMultiVerse Support Page</h2>
       </div>
+      <br/>
+      <br/>
+      <br/>
+      <div id="welP">
+          <h3> Our Support Page</h3>
+          <p>
+              Our goal is to get you back on track, with any issues you might be dealing with on your journey with us. <br/>
+              This pages contains most asked questions and some answers that we hope will clear up any problems. <br/>
+              Of course if you need more support or these answers are not clear enough for you, please find at the 
+              bottom of the  page some further <br/> contact details. You can email our helpful support team, 24/7 to get some help. 
+          </p>
+      </div>
+      <div id="sp1">
+          <h2 id>Forgotten my login details</h2>
+          <p>
+              For forgotten Login Details, please head to the login page, from there please locate forgotten password/username button.<br/>
+              You will receive an email that will allow you to update your login details with once complete will bring you to your login page<br/>
+              If you are still having any issues please use the contact information below to contact our team for more help.
+          </p>
+      </div> 
+      <div id="sp1">
+          <h2>How do I update my personal details</h2>
+          <p>
+              To update your personal details, please make sure to be login to your account. In your personal access page you will find an edit button. <br/>
+              Click the edit button, you will be brought to a edit screen and you will be able to change your details besides your subscription package.<br/>
+              If you are still having any issues please use the contact information below to contact our team for more help.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>Payment Methods we accept</h2>
+          <p>
+              Your convenience matters to us, and we want to ensure that you have the flexibility to choose the payment method that suits you best.<br/> 
+              Rest assured that we prioritize the security and privacy of your financial information, and all transactions are processed using the highest industry standards.<br/>
+              We accpet a multiple of cards, bank transfers, over the phone payments and more. If you are having any issues with these please do not hesitate to <br/>
+              use the contact information below to contact our team for more help.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2> Subscription Plans </h2>
+          <p>
+              To find out about our subscription plans, please go to our home page. From here you will be able to see all the details connected to each plan <br/>
+              Along with all the features you will have access to on each plan. A clear breakdon of pricing and types of payments can also be located on this page <br/>
+              If you are having any issues with these please do not hesitate to use the contact information below to contact our team for more help.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>Upgrade Subscription </h2>
+          <p>
+              To find out about our subscription plans, please go to our home page. From here you will be able to see all the details connected to each plan <br/>
+              If you wish to update your subscription to a higher or lower plan you will be able to change this in your account page. Please make sure that <br/>
+              you are login and can see your account details. From there you will locate an subscription button that request a statement to be entered of <br/>
+              what you are looking for which you will receive a copy in your emails and a case number. This will be checked by our customer support team <br/>
+              member. They will be in contact in 24 hours after the request has been sent. If more details are required they will contact you via email. <br/>
+              If you are having any issues with these please do not hesitate to use the contact information below to contact our team for more help.
+
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>Cancel Subscription</h2>
+          <p>
+              To cancel your subscription, please make to be login to your account and in your personal section you will <br/>
+              locate a cancel subscription button that will request a statement to be entered of what you are looking for which you will receive a copy <br/>
+              in your emails and a case number. This will be checked by our customer support team member. They will be in contact  <br/>
+              in 24 hours after the request has been sent. If more details are required they will contact you via email. <br/>
+              If you are having any issues with cancelling your subscription please use the contact information below to contact our team.
+          </p>
+      </div>
+
+      <div id="sp1">
+          <h2>How do I add Content to my list</h2>
+          <p>
+              To add content to your list, you must first locate the content that you wish to add. This can be done via are search function located on <br/>
+              our navigation bar. Once you located the content you want to add, click on the item and go into and from there you will see a button. <br/>
+              This button will be called add to list, once this is click the content will be added to your list. From there you will be able to access <br/>
+              your list feature and be able to see your content. Plese note this does depend on your subscription plan, which you can find more details <br/>
+              on our home page. If you are having any issues please use the contact information below to contact our team.
+          </p>
+      </div>
+
+      <div id="sp1">
+          <h2>How do I remove Content from my list</h2>
+          <p>
+              To remove content from your list, you first need to be login to your account. Once in your account you can locate the my list feature <br/>
+              This is depending on your subscription, details are found on our home page to see if you have this feature. On the navigation bar <br/>
+              click on the my list feature and select the content you wish to remove, once you are in that content you will see a button to remove from list. <br/>
+              Once you click this button the content will be removed and you will no longer see it. To re-add it to your list, please see instrustions above. <br/>
+              If you are having any issues please use the contact information below to contact our team.
+          </p>
+      </div>
+
+      <div id="sp1">
+          <h2>Having issues with some features</h2>
+          <p>
+              If you are having some issues with features on your account. This could be connected to you subscription plan. <br/>
+              Please check your account or our home page for subscription details. If you still can not see the issue please find contact details below. <br/>
+              For our team who will be able to help you in more detail, please make sure to have your account number on hand.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>How can I request a Film / Book / Tv Show to be added for viewing?</h2>
+          <p>
+              If there is some content that we dont have available and you would like to add in any of our forms. First please make sure that you are login <br/>
+              and that you are on the right subscription plan. Details for our subscriptions can be found on our home page. If you do have the subscription <br/>
+              But can not see the content you can send a request to our customer support team who will look into the possibility of getting the content <br/>
+              Please note that this can take a little while to gatther the request content but our customer support will keep you updated and you will <br/>
+              receive a case number that you can reference for any updates if you are having any queires to you request. please use the contact information <br/>
+               below to contact our team.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>Can I share my account with family / friends?</h2>
+          <p>
+              We do not have any issues with you sharing your account with your family and friends but please note that some features only appear with <br/>
+              certain subscriptions plans. This information can be found on our home page to see what subscription gives access to which features. <br/>
+              Please note that the list feature cannot be broken down into sub-list, and all subscriptions have only one accessable. <br/>
+              If you are having any issues please use the contact information below to contact our team.
+          </p>
+      </div>
+      <div id="sp1">
+          <h2>Can I access my account on multiple platforms?</h2>
+          <p>
+              This is correct, but please be aware of your subscription package as this controls what access to our features you have for more details. <br/>
+              Please find our answered question above or for further details please use the contact information below to contact our team for more help.
+          </p>
+      </div>
+
+
+      <br/><br/>
+      <div id="sp1-contact">
+          <h3>Contact Suppport Team for more help...</h3><br/>
+          <table>
+            <tr>
+              <th>Support Contact</th>
+              <th>Support Email</th>
+              <th>Community Board</th>
+            </tr>
+            <tr>
+               <td class="boader"> +353 020 915 8826 <br/>Time zone in Ireland (GMT+1)</td> 
+               <td>Technical Team: support.storyMV@gmail.com<br/><br/>Information Team: info.storyMV@gmail.com</td>
+               <td> Community Board details coming to a Verse near you!<br/> Please check Community Page for updates</td>
+            </tr>
+          </table>
+      </div>
+
+
+
       <FooterComponent />
     </div>
   );
@@ -22,3 +168,6 @@ export default function Support() {
 
 //------- Copyright & references ---------------
 //https://legacy.reactjs.org/docs/create-a-new-react-app.html
+
+
+// https://fakenumber.org/ireland